3785 - CountryRegionRepository.getManyRegionCountries
