+++ src/server/repository/assessmentCycle/countryRegion/index.ts
@@ -1,0 +1,5 @@
+import { getManyRegionCountries } from './getManyRegionCountries'
+
+export const CountryRegionRepository = {
+  getManyRegionCountries,
+}
